#10071: Merge silu_bw to TTNN
