+++ docs/source/ttnn/ttnn/ttnn/silu_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.silu_bw:
+
+  ttnn.silu_bw
+  #############
+
+  .. autofunction:: ttnn.silu_bw
