+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -285,11 +285,6 @@ std::vector<Tensor> rpow_bw(
     float exponent,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
-std::vector<Tensor> silu_bw(
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
-
 std::vector<Tensor> selu_bw(
     const Tensor& grad,
     const Tensor& input,
