+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -45,6 +45,7 @@ enum class UnaryBackwardOpType {
     LOG_BW,
     RELU6_BW,
     ABS_BW,
+    SILU_BW
 };
 
 struct UnaryBackwardFunction{
